changed place ship methods to player
