+++ lib/board.rb
@@ -21,20 +21,4 @@ class Board
     p row_5.values
   end
 
-
-  #
-  # def coord(x, y)
-  #   coord[y-1] = [@row_1[x], @row_2[x], @row_3[x], @row_4[x]] # y - 1 translates it into an index value
-  # end
-  #
-  #
-  # def y_coord
-  #   # row_1 = {1:"x", 2:"x", 3:"x"}
-  #   # row_2 = {1:"x", 2:"x", 3:"x"}
-  #   # row_3 = {1:"x", 2:"x", 3:"x"}
-  # end
-
-
-
-
 end
