+++ lib/game.rb
@@ -6,19 +6,19 @@ require 'pry'
 class Game
 
   def self.begin
-  # player_1 = Player.new('computer')
-  # player_2 = Player.new('human')
-  # Computer places the ships randomly
-  display
-  puts 'I have laid out my ships on the grid. You now need to layout your two ships.'
-  puts 'The first is two units long and the second is three units long.'
-  puts 'The grid has A1 at the top left and D4 at the bottom right.'
-  enter_coordinates
+    @board    = Board.new
+    @computer = Player.new('computer')
+    @human    = Player.new('human')
+    # computer do this method of placing ships randomly (@compouter.place_random)
+    display
+    puts 'I have laid out my ships on the grid. You now need to layout your two ships.'
+    puts 'The first is two units long and the second is three units long.'
+    puts 'The grid has A1 at the top left and D4 at the bottom right.'
+    enter_coordinates
   end
 
   def self.display
-  board = Board.new
-  board.display_board
+    @board.display_board
   end
 
   def self.enter_coordinates
