+++ lib/runner.rb
@@ -1,23 +1,23 @@
-require './place_ship'
-require './board'
-
-
-# ship1 = Place_ship.new(2)
-# ship2 = Place_ship.new(3)
+# require './place_ship'
+# require './board'
 #
-# p ship1
-# p ship2
-
-
-### SHOWS THAT NEW SHIPS CAN BE PLACED.
-computer_destroyer = Ship.new(2)
-computer_submarine = Ship.new(3)
-
-# c_destroyer_place = PlaceShip.new(computer_destroyer.health)
-# c_submarine_place = PlaceShip.new(computer_submarine.health)
-
-p c_destroyer_place
-p c_submarine_place
+#
+# # ship1 = Place_ship.new(2)
+# # ship2 = Place_ship.new(3)
+# #
+# # p ship1
+# # p ship2
+#
+#
+# ### SHOWS THAT NEW SHIPS CAN BE PLACED.
+# computer_destroyer = Ship.new(2)
+# computer_submarine = Ship.new(3)
+#
+# # c_destroyer_place = PlaceShip.new(computer_destroyer.health)
+# # c_submarine_place = PlaceShip.new(computer_submarine.health)
+# 
+# p c_destroyer_place
+# p c_submarine_place
 
 
 #
